requirements.txt
fastapi
uvicorn
requests==2.31.0
pytest
python-dotenv>=1.0.0